Average growth and stopping time are now implemented, only missing special cases with numbers that are way too big or specific numbers with too many steps to reach 1. (to be done later) Also, created the CollatzController class to work with scene builder and handle our events.

+++ CollatzApp/src/collatzapp/SequenceResult.java
@@ -16,10 +16,10 @@ public class SequenceResult {
     private final List<Integer> sequence;
     private final int stepsToOne;
     private final int peakNum;
-    private final int stopTime;
+    private final long stopTime;
     private final double avgGrowth;
     
-    public SequenceResult(int startNum, List<Integer> sequence, int stepsToOne, int peakNum, int stopTime, double avgGrowth) {
+    public SequenceResult(int startNum, List<Integer> sequence, int stepsToOne, int peakNum, long stopTime, double avgGrowth) {
         this.startNum = startNum;
         this.sequence = sequence;
         this.stepsToOne = stepsToOne;
@@ -45,7 +45,7 @@ public class SequenceResult {
         return stepsToOne;
     }
 
-    public int totalStoppingTime() { 
+    public long totalStoppingTime() { 
         return stopTime; 
     }
